+++ include/TGUI/AllWidgets.hpp
@@ -55,6 +55,7 @@
 #include <TGUI/Widgets/RangeSlider.hpp>
 #include <TGUI/Widgets/ScrollablePanel.hpp>
 #include <TGUI/Widgets/Scrollbar.hpp>
+#include <TGUI/Widgets/SeparatorLine.hpp>
 #include <TGUI/Widgets/Slider.hpp>
 #include <TGUI/Widgets/SpinButton.hpp>
 #include <TGUI/Widgets/SpinControl.hpp>